requirements.txt
streamlit==1.39.0
pandas==2.2.3
numpy==1.26.4
anthropic==0.39.0
python-dotenv==1.0.1
